backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
pydantic==1.10.12
SQLAlchemy==2.0.20
psycopg2-binary==2.9.6   # required only when using Postgres
requests==2.31.0
python-dotenv==1.0.0
PyJWT==2.8.0
rasa==3.7.0              # used only if running Rasa locally; can be adjusted
